+++ src/libs/Navigation/AppNavigator/BaseDrawerNavigator.js
@@ -6,7 +6,7 @@ import {View} from 'react-native';
 import styles, {getNavigationDrawerStyle, getNavigationDrawerType} from '../../../styles/styles';
 import withWindowDimensions, {windowDimensionsPropTypes} from '../../../components/withWindowDimensions';
 import Navigation from '../Navigation';
-import onlyWeb from '../../onlyWeb';
+import ignoreNative from '../../ignoreNative';
 
 const propTypes = {
     /** Screens to be passed in the Drawer */
@@ -34,14 +34,13 @@ const propTypes = {
 const Drawer = createDrawerNavigator();
 
 const BaseDrawerNavigator = (props) => {
-    const useLegacyImplementation = onlyWeb(() => props.isSmallScreenWidth);
     const content = (
         <Drawer.Navigator
             backBehavior="none"
             defaultStatus={Navigation.getDefaultDrawerState(props.isSmallScreenWidth)}
             sceneContainerStyle={styles.navigationSceneContainer}
             drawerContent={props.drawerContent}
-            useLegacyImplementation={useLegacyImplementation}
+            useLegacyImplementation={ignoreNative(props.isSmallScreenWidth)}
             screenOptions={{
                 cardStyle: styles.navigationScreenCardStyle,
                 headerShown: false,
