+++ src/libs/ignoreNative/index.js
@@ -1,0 +1,1 @@
+export default value => Boolean(value);
